Move TagButton component to separate file and update its css
