+++ front-end/src/BrowseRecipesPage.css
@@ -6,11 +6,6 @@
     margin-top: 10px;
 }
 
-.tagButton {
-    margin-right: 5px;
-    margin-bottom: 5px;
-}
-
 .recipeNameSearchbar {
     margin-bottom: 5px;
 }
